# verifiably-go/static/openapi.yaml
openapi: 3.0.3
info:
  title: Verifiably API
  version: 1.0.0
  description: |
    REST API for programmatic credential issuance, revocation, and OID4VP
    verification requests. All endpoints require an API key configured via the
    `VERIFIABLY_API_KEYS` environment variable (`name1:key1,name2:key2`).

    Each key is isolated in the issuance log under its own owner scope, so
    credentials issued by key A are never visible to key B.
servers:
  - url: /api/v1
    description: Current instance
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key obtained from the operator (VERIFIABLY_API_KEYS).
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: invalid or missing API key
...
security:
  - BearerAuth: []
paths:
...
  /schemas:
    get:
      operationId: listSchemas
      tags: [Catalog]
      summary: List custom schemas (owner-scoped)
      responses:
        "200": { description: Schemas., content: { application/json: { schema: { type: array, items: { type: object } } } } }
        "401": { description: Invalid or missing API key., content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    post:
      operationId: createSchema
      tags: [Catalog]
      summary: Create a custom credential schema
      requestBody:
        required: true
        content: { application/json: { schema: { type: object } } }
      responses:
        "201": { description: Schema created., content: { application/json: { schema: { type: object } } } }
        "400": { description: Bad request., content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "401": { description: Invalid or missing API key., content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...